Ability to draw polygon & handle polygon edit events

+++ googlemaps/src/main/java/com/vaadin/tapio/googlemaps/client/events/MapMoveListener.java
@@ -1,8 +1,8 @@
 package com.vaadin.tapio.googlemaps.client.events;
 
-import java.io.Serializable;
-
 import com.vaadin.tapio.googlemaps.client.LatLon;
+
+import java.io.Serializable;
 
 /**
  * Interface for listening map move and zoom events.
@@ -22,6 +22,6 @@ public interface MapMoveListener extends Serializable {
      * @param boundsSW
      *            The position of the south-west corner of the map.
      */
-    public void mapMoved(double zoomLevel, LatLon center, LatLon boundsNE,
+    public void mapMoved(int zoomLevel, LatLon center, LatLon boundsNE,
             LatLon boundsSW);
 }
